Automatically Generate a PDF and send it by Email
The challenge
Here, you are going to update the script cars.py. You will be using the above JSON data to process information. A part of the script is already done for you, where it calculates the car model with the most revenue (price * total_sales). You should now fulfil the following objectives with the script:

Calculate the car model which had the most sales.
a. Call format_car method for the car model.

Calculate the most popular car_year across all car make/models.